KAuthentication/backend/infrastructure/database/alembic/env.py
from logging.config import fileConfig
import asyncio
from sqlalchemy import pool
from sqlalchemy.ext.asyncio import create_async_engine
from alembic import context

# Import shared settings for database configuration
import sys
import os
from dotenv import load_dotenv

# Add project root to path (4 levels up from alembic dir)
backend_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../.."))
project_root = os.path.abspath(os.path.join(backend_dir, "../.."))
sys.path.insert(0, backend_dir)
sys.path.insert(0, project_root)

# Load .env file from project root
env_path = os.path.join(project_root, ".env")
load_dotenv(env_path)

from shared.backend.config.settings import settings

# Import models to register them with SQLModel metadata
from domain.utilities.BaseModel import BaseModels
from domain.client.models.UserProfileModels import UserModel
from domain.client.models.UserSessionModels import SessionModel

# this is the Alembic Config object, which provides
# access to the values within the .ini file in use.
config = context.config

# Override sqlalchemy.url with settings from environment
config.set_main_option("sqlalchemy.url", settings.database_url)

# Interpret the config file for Python logging.
# This line sets up loggers basically.
if config.config_file_name is not None:
    fileConfig(config.config_file_name)

# Set target metadata from SQLModel Base
target_metadata = BaseModels.metadata

# other values from the config, defined by the needs of env.py,
# can be acquired:
# my_important_option = config.get_main_option("my_important_option")
# ... etc.


def run_migrations_offline() -> None:
    """Run migrations in 'offline' mode.

    This configures the context with just a URL
    and not an Engine, though an Engine is acceptable
    here as well.  By skipping the Engine creation
    we don't even need a DBAPI to be available.

    Calls to context.execute() here emit the given string to the
    script output.

    """
    url = config.get_main_option("sqlalchemy.url")
    context.configure(
        url=url,
        target_metadata=target_metadata,
        literal_binds=True,
        dialect_opts={"paramstyle": "named"},
    )

    with context.begin_transaction():
        context.run_migrations()


async def run_async_migrations() -> None:
    """Run migrations in async mode using async engine."""
    connectable = create_async_engine(
        settings.database_url,
        poolclass=pool.NullPool,
    )

    async with connectable.connect() as connection:
        await connection.run_sync(do_run_migrations)

    await connectable.dispose()

def do_run_migrations(connection):
    """Helper function to run migrations in sync context."""
    context.configure(
        connection=connection,
        target_metadata=target_metadata
    )

    with context.begin_transaction():
        context.run_migrations()

def run_migrations_online() -> None:
    """Run migrations in 'online' mode with async support."""
    asyncio.run(run_async_migrations())

if context.is_offline_mode():
    run_migrations_offline()
else:
    run_migrations_online()
