KAuthentication/backend/pyproject.toml
[project]
name = "kauthentication-backend"
version = "0.0.1"
description = ""
authors = [
    {name = "matkumatmat",email = "mamattewahyu@gmail.com"}
]
license = {text = "mit"}
readme = "README.md"
packages = [
    {include = "application"},
    {include = "domain"},
    {include = "infrastructure"}
]

[tool.poetry.dependencies]
python = ">=3.14"
pydantic = ">=2.12.5,<3.0.0"
alembic = ">=1.18.2,<2.0.0"
fastapi = ">=0.128.0,<0.129.0"
redis = ">=7.1.0,<8.0.0"
pydantic-settings = ">=2.12.0,<3.0.0"
structlog = ">=25.5.0,<26.0.0"
asyncpg = ">=0.31.0,<0.32.0"
gunicorn = ">=24.1.1,<25.0.0"
ruff = ">=0.14.14,<0.15.0"
argon2-cffi = ">=23.1.0,<24.0.0"
pyjwt = ">=2.10.1,<3.0.0"
aiosmtplib = ">=3.0.0,<4.0.0"
uuid-utils = ">=0.9.0,<1.0.0"
mypy = ">=1.19.1,<2.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
uvicorn = ">=0.40.0,<0.41.0"
sqlalchemy = ">=2.0.0,<3.0.0"
pytest = "^9.0.2"
secure = "^1.0.1"
sqlmodel = "^0.0.32"
email-validator = "^2.3.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py314"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
    "TCH",
    "RUF",
]
ignore = [
    "E501",
    "B008",
    "B904",
    "N803",
    "I001",
    "UP043",
    "W293",
    "N802",
    "N999",
    "N815",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/**" = ["S101"]
"scripts/init_db.py" = ["F401"]
"**/persistence/*Model.py" = ["E711", "E712"]
"**/shared/Exceptions.py" = ["N818"]
"alembic/env.py" = ["E402", "F401", "I001"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "sqlalchemy.*",
    "alembic.*",
    "redis.*",
    "structlog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["tests_*.py", "*_tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
